refactor(#3879): bundle F — 3 files to tests/interfaces/
part of #3879

test_3300_p2a_queue_state_publish.py / test_chat_cli_flags.py /
test_textual_chat_copy_rewind_3362.py -- dominant_package() classifier
confirms interfaces subject for all 3 (lead-coder's bundle F ruling,
#3879 comment).

Two-direction audit: outward clean. Inward whole-repo grep found 2
STRUCTURAL imports (not prose, verified by reading each hit
individually rather than assuming pattern from prior bundles):
tests/runtime/test_text_effect_toggle_3796.py and
tests/interfaces/test_yank_failure_is_visible_3616.py both do
`from tests.test_textual_chat_copy_rewind_3362 import (...)` -- fixed
in this same commit to the new module path. Both are outside
test_textual_chat_copy_rewind_3362.py's OWN diff line, so
check_migration_diff_shape.py's pure-rename check does not gate them
(they are M lines on OTHER files, not on the renamed file itself).
Confirmed via pytest --collect-only: 19 tests collect cleanly after
the fix (2 were failing collection before).

Remaining inward hits (test_lazy_history_paging_3476.py,
test_copy_mode_3507.py, test_conversation_cursor_3476.py,
test_clipboard_pyperclip_3616.py, test_textual_chat_intervention_panel_3299.py)
are prose docstring/comment mentions, left stale per this migration
series' established discipline.
