+++ scripts/flat_tests_baseline.json
@@ -10,7 +10,6 @@
   "test_3126_doc_anchor_gate.py",
   "test_3233_inprocess_env_identity.py",
   "test_3237_worktree_clean_gate.py",
-  "test_3300_p2a_queue_state_publish.py",
   "test_3546_pipeline_driver_narrowing_inheritance.py",
   "test_3553_agent_step_worker_narrowing_inheritance.py",
   "test_3556_session_spawn_narrowing_inheritance.py",
@@ -18,7 +17,6 @@
   "test_3562_slash_session_new_narrowing_inheritance.py",
   "test_3723_flowview_pin_check.py",
   "test_3726_mypy_ratchet.py",
-  "test_3792_pr2_session_injection.py",
   "test_3794_p2_no_audit_event_prose.py",
   "test_3794_p3_audit_event_identifier_rename.py",
   "test_agent_construction_via_factory_997.py",
@@ -26,7 +24,6 @@
   "test_audit_events_single_assignment_3408.py",
   "test_builtin_skill_tool_name_drift_3092.py",
   "test_charter_citations_reference_live_content_3592.py",
-  "test_chat_cli_flags.py",
   "test_check_pr_closing_intent.py",
   "test_compaction_outcome_dist.py",
   "test_cost_chokepoint_ast_guard_1190.py",
@@ -70,7 +67,6 @@
   "test_safe_http_ssrf_redirect_1956.py",
   "test_safe_mcp_registry.py",
   "test_safe_process.py",
-  "test_sandbox_axis_contract_2983.py",
   "test_scheme_self_register_1608.py",
   "test_sp_prompt_package_coverage.py",
   "test_ssrf_guard_1956.py",
@@ -81,8 +77,6 @@
   "test_swe_bench_runner.py",
   "test_swe_bench_runner_chat_187.py",
   "test_swe_bench_runner_venv_183.py",
-  "test_template_env_2673.py",
-  "test_textual_chat_copy_rewind_3362.py",
   "test_tier_audit_format_pin.py",
   "test_tier_audit_private_state_ast.py",
   "test_tier_audit_source_segment_cache_3670.py",
