+++ tests/interfaces/test_chat_cli_flags.py
