+++ tests/interfaces/test_textual_chat_copy_rewind_3362.py
